Move tests into test/js/bun/import-attributes (not a regression)
